separated the dom from the api
