+++ .eslintrc.json
@@ -25,7 +25,8 @@
     "prefer-template":"off",
     "camelcase":"off",
     "class-methods-use-this":"off",
-    "no-nested-ternary":"off"
+    "no-nested-ternary":"off",
+    "import/no-cycle":"off"
   },
   "ignorePatterns": [
     "dist/",
